+++ src/indexes/__init__.py
@@ -1,4 +1,4 @@
-from .api import get_sp500, get_nasdaq100
+from .api import get_sp500, get_nasdaq100, get_sp100
 
-__version__ = "0.1.2"
-__all__ = ["get_sp500", "get_nasdaq100"]
+__version__ = "0.1.3"
+__all__ = ["get_sp500", "get_nasdaq100", "get_sp100"]